touchup: extract Details Basics Section
this seems like a good place for misc things to live.

makes notes feel more at-home/organized, and the graph part feels better
isolated as a header of the details.

included node-type-specific details in "basics" as well - maybe ideal
to have another section for these, but they aren't used very often,
so it seems ok.

also remove "problem node"/"addresses edge" text; seemed unnecessary.
